# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Life Kline Commercial API
  version: 1.0.0
  description: 商用 iOS AI 应用后端 API 契约。
servers:
  - url: http://127.0.0.1:3000
    description: Local development
security:
  - bearerAuth: []
paths:
...
  /api/auth/apple:
    post:
      summary: Production Sign in with Apple
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppleLoginPayload"
      responses:
        "200":
          description: JWT auth bundle.
        "401":
          description: Invalid Apple identity token.
        "422":
          description: Validation failed.
  /api/auth/wechat:
    post:
      summary: Production WeChat OAuth login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WeChatLoginPayload"
      responses:
        "200":
          description: JWT auth bundle.
        "401":
          description: WeChat authorization failed.
  /api/auth/refresh:
    post:
      summary: Rotate refresh token and issue a new access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        "200":
          description: New JWT auth bundle.
        "401":
          description: Refresh token expired or revoked.
  /api/auth/logout:
    post:
      summary: Revoke current session
      responses:
        "200":
          description: Logged out.
        "401":
          description: Authentication required.
...
  /api/gemini/generateContent:
    post:
      summary: Gemini non-streaming proxy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Gemini response.
        "429":
          description: AI quota reached.
        "503":
          description: AI provider not configured or unavailable.
  /api/gemini/generateContentStream:
    post:
      summary: Gemini SSE streaming proxy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: SSE stream.
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    cookieAuth:
      type: apiKey
      in: cookie
      name: life_session
  schemas:
...
    AppleLoginPayload:
      type: object
      required: [identityToken]
      properties:
        identityToken:
          type: string
        nonce:
          type: string
          maxLength: 256
        nonceSha256:
          type: string
          maxLength: 256
        fullName:
          type: string
          maxLength: 120
        email:
          type: string
          format: email
    WeChatLoginPayload:
      type: object
      required: [code]
      properties:
        code:
          type: string
          maxLength: 512
...